make context reducer a part of more general app state reducer
